:truck: rename to unconditional-image-generation
