+++ js/src/lib/components/Icons/IconUnconditionalImageGeneration.svelte
