+++ js/src/lib/interfaces/Types.ts
@@ -37,7 +37,7 @@ export enum PipelineType {
 	"text-to-image"                                           = "Text-to-Image",
 	"image-to-text"                                           = "Image-to-Text",
 	"image-to-image"                                          = "Image-to-Image",
-	"image-generation"                                        = "Image Generation",
+	"unconditional-image-generation"                          = "Unconditional Image Generation",
 	/// others
 	"structured-data-classification"                          = "Structured Data Classification",
 	"reinforcement-learning"                                  = "Reinforcement Learning",
@@ -103,7 +103,7 @@ export const PIPELINE_TAG_MODALITIES: Record<keyof typeof PipelineType, Modality
 	"text-to-image":                  "cv",
 	"image-to-text":                  "cv",
 	"image-to-image":                 "cv",
-	"image-generation":               "cv",
+	"unconditional-image-generation": "cv",
 	"structured-data-classification": "other",
 	"reinforcement-learning":         "other",
 };
@@ -125,7 +125,7 @@ export const PIPELINE_COLOR: {
 	"image-to-text":                  "red",
 	"object-detection":               "orange",
 	"image-to-image":                 "indigo",
-	"image-generation":               "green",
+	"unconditional-image-generation": "green",
 	"question-answering":             "blue",
 	"sentence-similarity":            "orange",
 	"structured-data-classification": "indigo",
@@ -174,7 +174,7 @@ export const PIPELINE_TAGS_DISPLAY_ORDER: Array<keyof typeof PipelineType> = [
 	"text-to-image",
 	"image-to-text",
 	"image-to-image",
-	"image-generation",
+	"unconditional-image-generation",
 	/// others
 	"structured-data-classification",
 	"reinforcement-learning",
