+++ tasks/src/const.ts
@@ -32,7 +32,7 @@ export const TASKS_MODEL_LIBRARIES: Record<keyof typeof PipelineType, Array<keyo
 	"text-to-image":                  [],
 	"image-to-text":                  [],
 	"image-to-image":                 [],
-	"image-generation":               [],
+	"unconditional-image-generation": [],
 	/// others
 	"structured-data-classification": ["sklearn"],
 	"reinforcement-learning":         ["transformers", "stable-baselines3"],
